LayoutTests/accessibility/custom-elements/current-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("custom-location").currentStateValue is "location"
PASS accessibilityController.accessibleElementById("custom-page").currentStateValue is "page"
PASS successfullyParsed is true

TEST COMPLETE
